chore(auth): refactor auth endpoints
Create a controllers folder and move all route handlers into an controller file in the controllers folder.
